+++ src/euphorie/client/browser/country.py
@@ -166,6 +166,12 @@ class SessionsView(BrowserView):
             type='session',
         )
 
+    def filter_valid_sessions(self, sessions):
+        return [
+            session for session in sessions
+            if self.get_survey_by_path(session.zodb_path)
+        ]
+
     @memoize
     def get_sessions(self):
         ''' Given some sessions create a tree
@@ -178,7 +184,8 @@ class SessionsView(BrowserView):
             sessions = self.account.sessions + self.account.acquired_sessions
         else:
             sessions = self.account.sessions
-        return sessions
+
+        return self.filter_valid_sessions(sessions)
 
     @memoize
     def get_ordered_sessions(self):
@@ -288,11 +295,16 @@ class SessionBrowserNavigator(SessionsView):
 
     no_splash = True
 
+    @property
+    @memoize
+    def groupid(self):
+        return self.request.get('groupid')
+
     @memoize
     def get_root_group(self):
         ''' Return the group that is the root of the navigation tree
         '''
-        groupid = self.request.get('groupid')
+        groupid = self.groupid
         if not groupid:
             return
         base_query = Session.query(Group).order_by(Group.short_name)