+++ docs/changes.rst
@@ -4,6 +4,7 @@ Changelog
 10.0.0b2 (unreleased)
 ---------------------
 
+- Upgrade to Plone 5.1.4
 - A new view ``@@refresh-resources-timestamp`` has been introduced
   to break the browser cache